+++ include/ThresExp.h
@@ -40,62 +40,62 @@ class ThresExp{
         virtual ~ThresExp();
 
         // Attribute that compute the expanded results
-        /* std::vector<std::complex<double>> ThresExpExpr( */
-        /*     std::complex<double> N, double xp */
+        /* std::vector<std::complex<long double>> ThresExpExpr( */
+        /*     std::complex<long double> N, long double xp */
         /* ); */
-        std::complex<double> ThresExpExpr(std::complex<double> N, double pt);
+        std::complex<long double> ThresExpExpr(std::complex<long double> N, long double pt);
 
         // Matching Coefficients
-        double Hth1gggH(double xp);
-        double Hth1gqqH(double xp);
-        double Hth1qqgH(double xp);
+        long double Hth1gggH(long double xp);
+        long double Hth1gqqH(long double xp);
+        long double Hth1qqgH(long double xp);
 
         // LO pt-distribution
-        std::complex<double> LOgggH(std::complex<double> N, double xp);
-        std::complex<double> LOgqqH(std::complex<double> N, double xp);
-        std::complex<double> LOqqgH(std::complex<double> N, double xp);
+        std::complex<long double> LOgggH(std::complex<long double> N, long double xp);
+        std::complex<long double> LOgqqH(std::complex<long double> N, long double xp);
+        std::complex<long double> LOqqgH(std::complex<long double> N, long double xp);
 
     private:
         int NC, NF, CA, ORD, CHANNEL;
-        double LF, LR, LQ;
-        double MH2, MUF2, MUR2, SROOT;
-        double CF, aass, SIGMA0;
+        long double LF, LR, LQ;
+        long double MH2, MUF2, MUR2, SROOT;
+        long double CF, aass, SIGMA0;
 
 
         // Beta functions
-        double Beta0;
-        double Beta1;
-        double Beta2;
+        long double Beta0;
+        long double Beta1;
+        long double Beta2;
 
         // Cusp Anomalous Dimensions
-        double Ath1g;
-        double Ath2g;
-        double Ath1q;
-        double Ath2q;
-        double Bth1g;
-        double Bth1q;
+        long double Ath1g;
+        long double Ath2g;
+        long double Ath1q;
+        long double Ath2q;
+        long double Bth1g;
+        long double Bth1q;
 
         // Sigma functions
         // gg->g
-        double Sigma22ggg(double xp);
-        double Sigma21ggg(double xp);
-        double Sigma20ggg(double xp);
+        long double Sigma22ggg(long double xp);
+        long double Sigma21ggg(long double xp);
+        long double Sigma20ggg(long double xp);
         // gq->g
-        double Sigma22gqg(double xp);
-        double Sigma21gqg(double xp);
-        double Sigma20gqg(double xp);
+        long double Sigma22gqg(long double xp);
+        long double Sigma21gqg(long double xp);
+        long double Sigma20gqg(long double xp);
         // qq->g
-        double Sigma22qqg(double xp);
-        double Sigma21qqg(double xp);
-        double Sigma20qqg(double xp);
+        long double Sigma22qqg(long double xp);
+        long double Sigma21qqg(long double xp);
+        long double Sigma20qqg(long double xp);
 
         // Zeta functions
-        const double zeta2=gsl_sf_zeta_int(2);
-        const double zeta3=gsl_sf_zeta_int(3);
-        const double zeta4=gsl_sf_zeta_int(4);
+        const long double zeta2=gsl_sf_zeta_int(2);
+        const long double zeta3=gsl_sf_zeta_int(3);
+        const long double zeta4=gsl_sf_zeta_int(4);
 
 
         // EulerGamma
         // TODO: Move this to global
-        const double EulerGamma=0.5772156649;
+        const long double EulerGamma=0.5772156649;
 };
